# openapi/openapi.yaml
openapi: 3.1.0
info:
    title: SoulLocker API
    version: 0.0.1
paths:
    /lobby:
        post:
            tags:
                - Lobby
            summary: Creates a new lobby
            description: Creates a new lobby for players to join and manage their soullink run
            operationId: create-lobby
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LobbyCreationRequest'
                required: true
            responses:
                "201":
                    description: Created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LobbyCreationResponse'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
    /lobby/{id}:
        get:
            tags:
                - Lobby
            summary: Retrieve a lobby
            description: Retrieve a lobby in which players manage their soullink run
            operationId: get-lobby
            parameters:
                - name: id
                  in: path
                  description: Lobby ID as a UUID
                  required: true
                  schema:
                    type: string
                    description: Lobby ID as a UUID
                    format: uuid
                    examples:
                        - 73beb67c-70c5-4c95-b99e-73e3c076f82f
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetLobbyResponse'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
components:
    schemas:
        ErrorDetail:
            type: object
            additionalProperties: false
            properties:
                location:
                    type: string
                    description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
                message:
                    type: string
                    description: Error message text
                value:
                    description: The value at the given location
        ErrorModel:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://example.com/schemas/ErrorModel.json
                    readOnly: true
                detail:
                    type: string
                    description: A human-readable explanation specific to this occurrence of the problem.
                    examples:
                        - Property foo is required but is missing.
                errors:
                    type: array
                    description: Optional list of individual error details
                    items:
                        $ref: '#/components/schemas/ErrorDetail'
                instance:
                    type: string
                    description: A URI reference that identifies the specific occurrence of the problem.
                    format: uri
                    examples:
                        - https://example.com/error-log/abc123
                status:
                    type: integer
                    description: HTTP status code
                    format: int64
                    examples:
                        - 400
                title:
                    type: string
                    description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
                    examples:
                        - Bad Request
                type:
                    type: string
                    description: A URI reference to human-readable documentation for the error.
                    format: uri
                    default: about:blank
                    examples:
                        - https://example.com/errors/example
        GetLobbyResponse:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://example.com/schemas/GetLobbyResponse.json
                    readOnly: true
                gameEditionId:
                    type: string
                id:
                    type: string
                name:
                    type: string
            required:
                - id
                - name
                - gameEditionId
        LobbyCreationRequest:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://example.com/schemas/LobbyCreationRequest.json
                    readOnly: true
                gameEditionId:
                    type: string
                    description: The game edition id of the pokemon version
                    examples:
                        - firered
                    minLength: 1
                    maxLength: 255
                name:
                    type: string
                    examples:
                        - Weekend Lobby
                    minLength: 10
                    maxLength: 255
                password:
                    type: string
                    examples:
                        - f00b4r1234
                    minLength: 8
                    maxLength: 255
            required:
                - name
                - password
                - gameEditionId
        LobbyCreationResponse:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://example.com/schemas/LobbyCreationResponse.json
                    readOnly: true
                lobbyId:
                    type: string
                    description: Lobby ID as a UUID
                    examples:
                        - 73beb67c-70c5-4c95-b99e-73e3c076f82f
            required:
                - lobbyId